requirements-test.txt
-r requirements.txt

pytest==6.2.3
pytest-django==4.2.0

# required by the assembly pipeline
requests~=2.21.0
# pip install -U git+git://github.com/EBI-Metagenomics/ena-api-handler.git